backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
requests==2.31.0
psycopg2-binary==2.9.3
Pillow==10.2.0
django-filter==23.5